+++ sdk/cognitivelanguage/Azure.AI.Language.QuestionAnswering/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Breaking Changes
 
-- Renamed `projectName` and `deploymentName` tags reported on `QuestionAnsweringClient` activities to `az.question_answering.project.name` and `az.question_answering.deployment.name` following OpenTelemetry attribute naming conventions.
+- Renamed `projectName` and `deploymentName` tags reported on `QuestionAnsweringClient` activities to `az.cognitivelanguage.project.name` and `az.cognitivelanguage.deployment.name` following OpenTelemetry attribute naming conventions.
 
 ### Bugs Fixed